% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Must be a data frame with three variables (v1, v2, w)}

\item{init_node}{Must be numeric scalar or integer}
}
\value{
The shortest path from the
initial node to every other node
}
\description{
\code{dijkstra} takes a graph and an initial node and calculates the shortest path from the
initial node to every other node in the graph.
}
\details{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)


}
