% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc.R
\name{euc}
\alias{euc}
\title{Euclidean Algorithm}
\usage{
euc(a, b)
}
\arguments{
\item{a}{Must be numeric scalar or integer}

\item{b}{Must be numeric scalar or integer}
}
\value{
The greatest common divisor of \code{a} and \code{b}
}
\description{
The goal of this algorithm is to find the greatest common divisor of two input (a,b) numbers
}
\examples{
euc(100,10)

\dontrun{
euc('dog','cat')
}

}
