% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Must be numeric scalar or integer}

\item{b}{Must be numeric scalar or integer}
}
\value{
The greatest common divisor of \code{a} and \code{b}
}
\description{
\code{euclidean} returns the greatest common divisor of two input (a,b) values.
}
\details{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(100,10)

\dontrun{
euclidean('dog','cat')
}




}
