% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03.R
\docType{package}
\name{lab03}
\alias{lab03}
\title{Package with Euclidean and Dijkstra's algorithms}
\description{
Two functions are included:
\code{euclidean} and \code{dijkstra}
}
\details{
For further information check functions' documentation.
}
\seealso{
Useful links:

\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}

\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\author{
\itemize{
\item Farid Musayev
\item Mucahit Sahin
\item Selen Karaduman
}
}
