% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Dataset for Dijkstra Algorithm}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{1st vertex}
\item{v2}{2nd vertex}
\item{w}{weight between two vertices}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset contains corresponding vertices and their weights.
}
\keyword{datasets}
